requirements.txt
numpy==1.22.0
scikit-learn==1.5.0
streamlit==1.36.0
streamlit-option-menu==0.3.13
pickle-mixin==1.0.2
pyrebase4==4.10.0
